requirements.txt
-i https://pypi.org/simple
bokeh==2.0.2
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
dask[complete]==2.15.0
decorator==4.4.2
distributed==2.15.0
fsspec==0.7.3
gcsfs==0.6.1
google-api-core==1.17.0
google-auth-oauthlib==0.4.1
google-auth==1.14.1
google-cloud-core==1.3.0
google-cloud-storage==1.28.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
heapdict==1.0.1
idna==2.9
jinja2==3.0.0a1
locket==0.2.0
markupsafe==2.0.0a1
msgpack==1.0.0
numpy==1.18.3
oauthlib==3.1.0
packaging==20.3
pandas==1.0.3
partd==1.1.0
pillow==7.1.2
protobuf==3.11.3
psutil==5.7.0
pyarrow==0.17.0
pyasn1-modules==0.2.8
pyasn1==0.4.8
pyparsing==3.0.0a1
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2019.3
pyyaml==5.3.1
requests-oauthlib==1.3.0
requests==2.23.0
rsa==4.0
six==1.14.0
sortedcontainers==2.1.0
tblib==1.6.0
toolz==0.10.0
tornado==6.0.4 ; python_version < '3.8'
typing-extensions==3.7.4.2
urllib3==1.25.9
zict==2.0.0
